+++ pkg/sys/system.go
@@ -37,6 +37,7 @@ import (
 	"google.golang.org/protobuf/proto"
 
 	futuapi "github.com/shing1211/futuapi4go/internal/client"
+	"github.com/shing1211/futuapi4go/pkg/constant"
 	"github.com/shing1211/futuapi4go/pkg/pb/common"
 	"github.com/shing1211/futuapi4go/pkg/pb/getdelaystatistics"
 	"github.com/shing1211/futuapi4go/pkg/pb/getglobalstate"
@@ -46,7 +47,19 @@ import (
 
 // wrapError standardizes error messages for proto response failures
 func wrapError(funcName string, retType int32, retMsg string) error {
-	return fmt.Errorf("%s failed: retType=%d, retMsg=%s", funcName, retType, retMsg)
+	code := constant.ErrorCode(retType)
+	if retType == 0 {
+		code = constant.ErrCodeSuccess
+	} else if retType < 0 {
+		code = constant.ErrorCode(retType)
+	} else {
+		code = constant.ErrCodeUnknown
+	}
+	return &constant.FutuError{
+		Code:    code,
+		Message: retMsg,
+		Func:   funcName,
+	}
 }
 
 const (